+++ lib/posix/options/Kconfig
@@ -30,6 +30,7 @@ rsource "Kconfig.confstr"
 rsource "Kconfig.env"
 rsource "Kconfig.eventfd"
 rsource "Kconfig.fdtable"
+rsource "Kconfig.fd_mgmt"
 rsource "Kconfig.fnmatch"
 rsource "Kconfig.fs"
 rsource "Kconfig.getentropy"
